Split test_cache.py into modules matching the modules it tests.
The file was way too big.
