+++ src/relstorage/cache/tests/test_storage_cache.py
@@ -1,0 +1,426 @@
+##############################################################################
+#
+# Copyright (c) 2009 Zope Foundation and Contributors.
+# All Rights Reserved.
+#
+# This software is subject to the provisions of the Zope Public License,
+# Version 2.1 (ZPL).  A copy of the ZPL should accompany this distribution.
+# THIS SOFTWARE IS PROVIDED "AS IS" AND ANY AND ALL EXPRESS OR IMPLIED
+# WARRANTIES ARE DISCLAIMED, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+# WARRANTIES OF TITLE, MERCHANTABILITY, AGAINST INFRINGEMENT, AND FITNESS
+# FOR A PARTICULAR PURPOSE.
+#
+##############################################################################
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+
+
+import threading
+
+from ZODB.utils import p64
+
+from relstorage.tests import TestCase
+
+from . import MockOptionsWithFakeMemcache as MockOptionsWithFakeCache
+from . import MockAdapter
+
+class StorageCacheTests(TestCase):
+    # pylint:disable=too-many-public-methods
+
+    def setUp(self):
+        from relstorage.tests.fakecache import data
+        data.clear()
+        self._instances = []
+
+    def tearDown(self):
+        from relstorage.tests.fakecache import data
+        data.clear()
+        for inst in self._instances:
+            inst.close(close_async=False)
+
+    def getClass(self):
+        from relstorage.cache import StorageCache
+        return StorageCache
+
+    def _makeOne(self, **kwargs):
+        options = MockOptionsWithFakeCache.from_args(**kwargs)
+        inst = self.getClass()(MockAdapter(), options,
+                               'myprefix')
+        self._instances.append(inst)
+        return inst.new_instance() # coverage and sharing testing
+
+    def test_ctor(self):
+        from relstorage.tests.fakecache import Client
+        from relstorage.cache.storage_cache import _MemcacheStateCache
+        c = self._makeOne()
+        self.assertEqual(len(c.clients_local_first), 2)
+        self.assertEqual(len(c.clients_global_first), 2)
+        self.assertIsInstance(c.clients_global_first[0], _MemcacheStateCache)
+        self.assertIsInstance(c.clients_global_first[0].client, Client)
+        self.assertEqual(c.clients_global_first[0].client.servers, ['host:9999'])
+        self.assertEqual(c.prefix, 'myprefix')
+
+        # can be closed multiple times
+        c.close()
+        c.close()
+        self.test_closed_state(c)
+
+    def test_closed_state(self, c=None):
+        if c is None:
+            c = self._makeOne()
+        c.close(close_async=False)
+
+        self.assertEqual(len(c), 0)
+        self.assertTrue(c)
+        self.assertEqual(c.size, 0)
+        self.assertEqual(c.limit, 0)
+        # At no point did we spawn threads.
+        self.assertEqual(threading.active_count(), 1)
+
+    def test_stats(self):
+        inst = self._makeOne()
+        self.assertIsInstance(inst.stats(), dict)
+        inst.close()
+        self.assertIsInstance(inst.stats(), dict)
+
+    def test_save(self):
+        import tempfile
+        import os
+        import shutil
+
+        c = self._makeOne()
+        c.checkpoints = (0, 0)
+        c.tpc_begin()
+        c.store_temp(2, b'abc')
+        # tid is 2016-09-29 11:35:58,120
+        tid = 268595726030645777
+        c.after_tpc_finish(p64(268595726030645777))
+
+        key = list(iter(c.local_client))[0]
+        self.assertEqual((2, tid), key)
+
+        c.options.cache_local_dir = tempfile.mkdtemp()
+        self.addCleanup(shutil.rmtree, c.options.cache_local_dir, True)
+
+        c.save()
+        files = os.listdir(c.options.cache_local_dir)
+        __traceback_info__ = files
+        # Older versions of sqlite may leave -shm and -wal
+        # files around.
+        self.assertGreaterEqual(len(files), 1)
+
+        # Creating one in the same place automatically loads it.
+        c2 = self._makeOne(cache_local_dir=c.options.cache_local_dir)
+        self.assertEqual(1, len(c2))
+        self.test_closed_state(c2)
+        self.test_closed_state(c)
+
+    def test_clear(self):
+        from relstorage.tests.fakecache import data
+        data.clear()
+        c = self._makeOne()
+        data['x'] = '1'
+        c.clear()
+        self.assertFalse(data)
+        self.assertEqual(c.checkpoints, None)
+        self.assertEqual(dict(c.delta_after0), {})
+        self.assertEqual(dict(c.delta_after1), {})
+
+    def test_load_without_checkpoints(self):
+        c = self._makeOne()
+        res = c.load(None, 2)
+        self.assertEqual(res, (None, None))
+
+    def test_load_using_delta_after0_hit(self):
+        from relstorage.tests.fakecache import data
+        adapter = MockAdapter()
+        c = self.getClass()(adapter, MockOptionsWithFakeCache(), 'myprefix')
+        c.current_tid = 60
+        c.checkpoints = (50, 40)
+        c.delta_after0[2] = 55
+        data['myprefix:state:55:2'] = p64(55) + b'abc'
+        res = c.load(None, 2)
+        self.assertEqual(res, (b'abc', 55))
+
+    def test_load_using_delta_after0_miss(self):
+        adapter = MockAdapter()
+        c = self.getClass()(adapter, MockOptionsWithFakeCache(), 'myprefix')
+        c.current_tid = 60
+        c.checkpoints = (50, 40)
+        c.delta_after0[2] = 55
+        adapter.mover.data[2] = (b'abc', 55)
+        res = c.load(None, 2)
+        self.assertEqual(res, (b'abc', 55))
+
+    def test_load_using_delta_after0_inconsistent(self):
+        adapter = MockAdapter()
+        c = self.getClass()(adapter, MockOptionsWithFakeCache(), 'myprefix')
+        c.current_tid = 60
+        c.checkpoints = (50, 40)
+        c.delta_after0[2] = 55
+        adapter.mover.data[2] = (b'abc', 56)
+        with self.assertRaisesRegex(AssertionError, "Detected an inconsistency"):
+            c.load(None, 2)
+
+    def test_load_using_delta_after0_future_error(self):
+        adapter = MockAdapter()
+        c = self.getClass()(adapter, MockOptionsWithFakeCache(), 'myprefix')
+        c.current_tid = 55
+        c.checkpoints = (50, 40)
+        c.delta_after0[2] = 55
+        adapter.mover.data[2] = ('abc', 56)
+        from ZODB.POSException import ReadConflictError
+        with self.assertRaisesRegex(ReadConflictError, "future"):
+            c.load(None, 2)
+
+    def test_load_using_checkpoint0_hit(self):
+        from relstorage.tests.fakecache import data
+        adapter = MockAdapter()
+        c = self.getClass()(adapter, MockOptionsWithFakeCache(), 'myprefix')
+        c.current_tid = 60
+        c.checkpoints = (50, 40)
+        data['myprefix:state:50:2'] = p64(45) + b'xyz'
+        res = c.load(None, 2)
+        self.assertEqual(res, (b'xyz', 45))
+
+    def test_load_using_checkpoint0_miss(self):
+        from relstorage.tests.fakecache import data
+        adapter = MockAdapter()
+        c = self.getClass()(adapter, MockOptionsWithFakeCache(), 'myprefix')
+        c.current_tid = 60
+        c.checkpoints = (50, 40)
+        adapter.mover.data[2] = (b'xyz', 45)
+        res = c.load(None, 2)
+        self.assertEqual(res, (b'xyz', 45))
+        self.assertEqual(data.get('myprefix:state:50:2'), p64(45) + b'xyz')
+
+    def test_load_using_delta_after1_hit(self):
+        from relstorage.tests.fakecache import data
+        adapter = MockAdapter()
+        c = self.getClass()(adapter, MockOptionsWithFakeCache(), 'myprefix')
+        c.current_tid = 60
+        c.checkpoints = (50, 40)
+        c.delta_after1[2] = 45
+        data['myprefix:state:45:2'] = p64(45) + b'abc'
+        res = c.load(None, 2)
+        self.assertEqual(res, (b'abc', 45))
+        self.assertEqual(data.get('myprefix:state:50:2'), p64(45) + b'abc')
+
+    def test_load_using_delta_after1_miss(self):
+        from relstorage.tests.fakecache import data
+        adapter = MockAdapter()
+        c = self.getClass()(adapter, MockOptionsWithFakeCache(), 'myprefix')
+        c.current_tid = 60
+        c.checkpoints = (50, 40)
+        c.delta_after1[2] = 45
+        adapter.mover.data[2] = (b'abc', 45)
+        res = c.load(None, 2)
+        self.assertEqual(res, (b'abc', 45))
+        self.assertEqual(data.get('myprefix:state:50:2'), p64(45) + b'abc')
+
+    def test_load_using_checkpoint1_hit(self):
+        from relstorage.tests.fakecache import data
+        __traceback_info__ = data
+        adapter = MockAdapter()
+        c = self.getClass()(adapter, MockOptionsWithFakeCache(), 'myprefix')
+        c.current_tid = 60
+        c.checkpoints = (50, 40)
+        data['myprefix:state:40:2'] = p64(35) + b'123'
+        res = c.load(None, 2)
+        self.assertEqual(res, (b'123', 35))
+        self.assertEqual(data.get('myprefix:state:50:2'), p64(35) + b'123')
+
+    def test_load_using_checkpoint1_miss(self):
+        from relstorage.tests.fakecache import data
+        adapter = MockAdapter()
+        c = self.getClass()(adapter, MockOptionsWithFakeCache(), 'myprefix')
+        c.current_tid = 60
+        c.checkpoints = (50, 40)
+        adapter.mover.data[2] = (b'123', 35)
+        res = c.load(None, 2)
+        self.assertEqual(res, (b'123', 35))
+        self.assertEqual(data.get('myprefix:state:50:2'), p64(35) + b'123')
+
+    def test_store_temp(self):
+        c = self._makeOne()
+        c.tpc_begin()
+        c.store_temp(2, b'abc')
+        c.store_temp(1, b'def')
+        c.store_temp(2, b'ghi')
+        self.assertEqual(b'ghi', c.read_temp(2))
+        self.assertEqual(c.queue_contents, {1: (3, 6), 2: (6, 9)})
+        c.queue.seek(0)
+        self.assertEqual(c.queue.read(), b'abcdefghi')
+        c.checkpoints = (1, 0)
+        c.after_tpc_finish(p64(3))
+
+        self.assertEqual(dict(c.delta_after0), {2: 3, 1: 3})
+
+    def test_send_queue_small(self):
+        from relstorage.tests.fakecache import data
+        c = self._makeOne()
+        c.tpc_begin()
+        c.store_temp(2, b'abc')
+        c.store_temp(3, b'def')
+        tid = p64(55)
+        c.send_queue(tid)
+        self.assertEqual(data, {
+            'myprefix:state:55:2': tid + b'abc',
+            'myprefix:state:55:3': tid + b'def',
+            })
+        self.assertEqual(len(c), 2)
+
+    def test_send_queue_large(self):
+        from relstorage.tests.fakecache import data
+        c = self._makeOne()
+        c.send_limit = 100
+        c.tpc_begin()
+        c.store_temp(2, b'abc')
+        c.store_temp(3, b'def' * 100)
+        tid = p64(55)
+        c.send_queue(tid)
+        self.assertEqual(data, {
+            'myprefix:state:55:2': tid + b'abc',
+            'myprefix:state:55:3': tid + (b'def' * 100),
+            })
+
+    def test_send_queue_none(self):
+        from relstorage.tests.fakecache import data
+        c = self._makeOne()
+        c.tpc_begin()
+        tid = p64(55)
+        c.send_queue(tid)
+        self.assertEqual(data, {})
+
+    def test_after_tpc_finish(self):
+        c = self._makeOne()
+        c.tpc_begin()
+        c.after_tpc_finish(p64(55))
+        c.after_tpc_finish(p64(55))
+        # XXX: This test doesn't actually assert anything. It used to check
+        # the commit-count key, but we don't use that anymore.
+
+    def test_clear_temp(self):
+        c = self._makeOne()
+        c.tpc_begin()
+        c.clear_temp()
+        self.assertEqual(c.queue_contents, None)
+        self.assertEqual(c.queue, None)
+
+    def test_after_poll_init_checkpoints(self):
+        from relstorage.tests.fakecache import data
+        c = self._makeOne()
+        c.after_poll(None, 40, 50, [])
+        self.assertEqual(c.checkpoints, (50, 50))
+        self.assertEqual(data['myprefix:checkpoints'], b'50 50')
+
+    def test_after_poll_ignore_garbage_checkpoints(self):
+        from relstorage.tests.fakecache import data
+        data['myprefix:checkpoints'] = 'baddata'
+        c = self._makeOne()
+        c.after_poll(None, 40, 50, [])
+        self.assertEqual(c.checkpoints, (50, 50))
+        self.assertEqual(data['myprefix:checkpoints'], b'50 50')
+
+    def test_after_poll_ignore_invalid_checkpoints(self):
+        from relstorage.tests.fakecache import data
+        data['myprefix:checkpoints'] = b'60 70'  # bad: c0 < c1
+        c = self._makeOne()
+        c.after_poll(None, 40, 50, [])
+        self.assertEqual(c.checkpoints, (50, 50))
+        self.assertEqual(data['myprefix:checkpoints'], b'50 50')
+
+    def test_after_poll_reinstate_checkpoints(self):
+        from relstorage.tests.fakecache import data
+        self.assertEqual(data, {})
+        c = self._makeOne()
+        c.checkpoints = (40, 30)
+        c.after_poll(None, 40, 50, [])
+        self.assertEqual(c.checkpoints, (50, 50))
+        self.assertEqual(data['myprefix:checkpoints'], b'40 30')
+
+    def test_after_poll_future_checkpoints_when_cp_exist(self):
+        from relstorage.tests.fakecache import data
+        data['myprefix:checkpoints'] = b'90 80'
+        c = self._makeOne()
+        c.checkpoints = (40, 30)
+        c.current_tid = 40
+        c.after_poll(None, 40, 50, [(2, 45)])
+        # This instance can't yet see txn 90, so it sticks with
+        # the existing checkpoints.
+        self.assertEqual(c.checkpoints, (40, 30))
+        self.assertEqual(data['myprefix:checkpoints'], b'90 80')
+        self.assertEqual(dict(c.delta_after0), {2: 45})
+        self.assertEqual(dict(c.delta_after1), {})
+
+    def test_after_poll_future_checkpoints_when_cp_nonexistent(self):
+        from relstorage.tests.fakecache import data
+        data['myprefix:checkpoints'] = b'90 80'
+        c = self._makeOne()
+        c.after_poll(None, 40, 50, [(2, 45)])
+        # This instance can't yet see txn 90, and there aren't any
+        # existing checkpoints, so fall back to the current tid.
+        self.assertEqual(c.checkpoints, (50, 50))
+        self.assertEqual(data['myprefix:checkpoints'], b'90 80')
+        self.assertEqual(dict(c.delta_after0), {})
+        self.assertEqual(dict(c.delta_after1), {})
+
+    def test_after_poll_retain_checkpoints(self):
+        from relstorage.tests.fakecache import data
+        data['myprefix:checkpoints'] = b'40 30'
+        c = self._makeOne()
+        c.checkpoints = (40, 30)
+        c.current_tid = 40
+        c.delta_after1 = {1: 35}
+        c.after_poll(None, 40, 50, [(2, 45), (2, 41)])
+        self.assertEqual(c.checkpoints, (40, 30))
+        self.assertEqual(data['myprefix:checkpoints'], b'40 30')
+        self.assertEqual(dict(c.delta_after0), {2: 45})
+        self.assertEqual(dict(c.delta_after1), {1: 35})
+
+    def test_after_poll_new_checkpoints(self):
+        from relstorage.tests.fakecache import data
+        data['myprefix:checkpoints'] = b'50 40'
+        adapter = MockAdapter()
+        c = self.getClass()(adapter, MockOptionsWithFakeCache(), 'myprefix')
+        # Note that OID 3 changed twice.  list_changes is not required
+        # to provide the list of changes in order, so simulate
+        # a list of changes that is out of order.
+        adapter.poller.changes = [(3, 42), (1, 35), (2, 45), (3, 41)]
+        c.checkpoints = (40, 30)
+        c.current_tid = 40
+        c.after_poll(None, 40, 50, [(3, 42), (2, 45), (3, 41)])
+        self.assertEqual(c.checkpoints, (50, 40))
+        self.assertEqual(data['myprefix:checkpoints'], b'50 40')
+        self.assertEqual(dict(c.delta_after0), {})
+        self.assertEqual(dict(c.delta_after1), {2: 45, 3: 42})
+
+    def test_after_poll_gap(self):
+        from relstorage.tests.fakecache import data
+        data['myprefix:checkpoints'] = b'40 30'
+        adapter = MockAdapter()
+        c = self.getClass()(adapter, MockOptionsWithFakeCache(), 'myprefix')
+        adapter.poller.changes = [(3, 42), (1, 35), (2, 45)]
+        c.checkpoints = (40, 30)
+        c.current_tid = 40
+        # provide a prev_tid_int that shows a gap in the polled
+        # transaction list, forcing a rebuild of delta_after(0|1).
+        c.after_poll(None, 43, 50, [(2, 45)])
+        self.assertEqual(c.checkpoints, (40, 30))
+        self.assertEqual(data['myprefix:checkpoints'], b'40 30')
+        self.assertEqual(dict(c.delta_after0), {2: 45, 3: 42})
+        self.assertEqual(dict(c.delta_after1), {1: 35})
+
+    def test_after_poll_shift_checkpoints(self):
+        from relstorage.tests.fakecache import data
+        data['myprefix:checkpoints'] = b'40 30'
+        c = self._makeOne()
+        c.delta_size_limit = 2
+        c.checkpoints = (40, 30)
+        c.current_tid = 40
+        c.after_poll(None, 40, 314, [(1, 45), (2, 46)])
+        self.assertEqual(c.checkpoints, (40, 30))
+        self.assertEqual(data['myprefix:checkpoints'], b'314 40')
+        self.assertEqual(dict(c.delta_after0), {1: 45, 2: 46})
+        self.assertEqual(dict(c.delta_after1), {})
